refactor : CollectionModule에서 PlaceModule에 대한 종속성 분리 - 기존에 크롤링을 통해 장소와 링크를 생성하는 로직을 Crawling Module로 모두 이관, CollectionModule과 PlaceModule은 각각에 대한 책임만 지도록 함.
