+++ src/collection/collection.service.ts
@@ -7,6 +7,7 @@ import { CollectionPlacesListResDto } from './dtos/collection-places-list-res.dt
 import { Transactional } from 'typeorm-transactional';
 import { CollectionsListResDto } from './dtos/paginated-collections-list-res.dto';
 import { throwIeumException } from 'src/common/utils/exception.util';
+import { Collection } from './entities/collection.entity';
 
 @Injectable()
 export class CollectionService {
@@ -14,46 +15,7 @@ export class CollectionService {
   constructor(
     private readonly collectionRepository: CollectionRepository,
     private readonly collectionPlaceRepository: CollectionPlaceRepository,
-    private readonly placeService: PlaceService,
   ) {}
-
-  @Transactional() //Transactional의 정상적인 작동 의심.
-  async createCollection(createCollectionReq: CreateCollectionReqDto) {
-    try {
-      if (
-        //중복이면 그대로 리턴해주는게 맞지 않은가? 고민해보기
-        await this.collectionRepository.isDuplicatedCollection(
-          createCollectionReq.userId,
-          createCollectionReq.link,
-        )
-      ) {
-        throwIeumException('CONFLICTED_COLLECTION');
-      }
-
-      const collection = await this.collectionRepository.createCollection(
-        createCollectionReq.userId,
-        createCollectionReq.collectionType,
-        createCollectionReq.link,
-        createCollectionReq.content,
-      );
-
-      await Promise.all(
-        createCollectionReq.placeKeywords.map(async (placeKeyword) => {
-          const place =
-            await this.placeService.createPlaceByKakaoLocal(placeKeyword);
-          await this.collectionPlaceRepository.createCollectionPlace(
-            collection.id,
-            place.id,
-            placeKeyword,
-          );
-        }),
-      );
-      return collection;
-    } catch (e) {
-      this.logger.error(e.message, e.stack);
-      throwIeumException('CREATE_COLLECTION_FAILED');
-    }
-  }
 
   async getUnviewedCollections(
     userId: number,
@@ -86,5 +48,40 @@ export class CollectionService {
       await this.collectionPlaceRepository.getCollectionPlaces(collectionId);
     await this.collectionRepository.updateIsViewed(userId, collectionId);
     return new CollectionPlacesListResDto(collectionPlaces, collectionId);
+  }
+
+  async createCollectionWithDuplicationCheck(
+    createCollectionReq: CreateCollectionReqDto,
+  ): Promise<Collection> {
+    if (
+      //중복이면 그대로 리턴해주는게 맞지 않은가? 고민해보기
+      await this.collectionRepository.isDuplicatedCollection(
+        createCollectionReq.userId,
+        createCollectionReq.link,
+      )
+    ) {
+      throwIeumException('CONFLICTED_COLLECTION');
+    }
+
+    const collection = await this.collectionRepository.createCollection(
+      createCollectionReq.userId,
+      createCollectionReq.collectionType,
+      createCollectionReq.link,
+      createCollectionReq.content,
+    );
+
+    return collection;
+  }
+
+  async createCollectionPlace(
+    collectionId: number,
+    placeId: number,
+    placeKeyword: string,
+  ) {
+    await this.collectionPlaceRepository.createCollectionPlace(
+      collectionId,
+      placeId,
+      placeKeyword,
+    );
   }
 }
